Add Aviary areas
