+++ src/assets/data/realms.json
@@ -7,7 +7,12 @@
       "imageUrl": "/assets/game/realms/Aviary.webp",
       "areas": [
         "Mhbh0g9OdG",
-        "tP8ZaD5Qwf"
+        "H2FgLkSRQw",
+        "JdoOxvcKlP",
+        "TUyoacebPE",
+        "rTIqNgvA4N",
+        "tP8ZaD5Qwf",
+        "pNNbszCpyA"
       ],
       "mapData": {
         "position": [-453, 364.75],
